sandbox_manager/requirements.txt
fastapi[standard]
docker
pydantic-settings
